Refactor hunk-git snapshot internals and commit path
Move normal commit creation to in-process git2 while keeping Git CLI fallback for signing-sensitive repos to avoid behavior regressions. Split oversized git snapshot internals into focused modules, document the intentional detailed-path behavior for `*_without_refresh` line stats/snapshots, and add a regression test to lock that contract.
